The number of sent messages is below or equal to the number of MPI ranks.

+++ code/VerticesExtractor.cpp
@@ -31,7 +31,6 @@
 void VerticesExtractor::process(int*m_mode_send_vertices_sequence_id,
 				ArrayOfReads*m_myReads,
 				bool*m_reverseComplementVertex,
-				int*m_mode_send_vertices_sequence_id_position,
 				int rank,
 				StaticVector*m_outbox,
 				bool*m_mode_send_vertices,
@@ -40,14 +39,27 @@ void VerticesExtractor::process(int*m_mode_send_vertices_sequence_id,
 				RingAllocator*m_outboxAllocator,
 				bool m_colorSpaceMode,int*m_mode
 				){
+	if(this->m_outbox==NULL){
+		m_rank=rank;
+		this->m_mode=m_mode;
+		this->m_outbox=m_outbox;
+		this->m_outboxAllocator=m_outboxAllocator;
+	}
 	#ifdef ASSERT
 	assert(m_pendingMessages>=0);
 	#endif
 	if(m_pendingMessages!=0){
 		return;
 	}
-	#ifdef SHOW_PROGRESS
-	if(*m_mode_send_vertices_sequence_id%100000==0 and *m_mode_send_vertices_sequence_id_position==0){
+
+	if(mustTriggerReduction()&&m_mode_send_vertices_sequence_id_position==0){
+		flushAll(m_outboxAllocator,m_outbox,rank);
+		checkPendingMessagesForReduction(m_outbox,rank);
+
+		return;
+	}
+	
+	if(*m_mode_send_vertices_sequence_id%100000==0 and m_mode_send_vertices_sequence_id_position==0){
 		string reverse="";
 		if(*m_reverseComplementVertex==true){
 			reverse="(reverse complement) ";
@@ -55,12 +67,19 @@ void VerticesExtractor::process(int*m_mode_send_vertices_sequence_id,
 		printf("Rank %i is computing vertices & edges %s[%i/%i]\n",rank,reverse.c_str(),(int)*m_mode_send_vertices_sequence_id+1,(int)m_myReads->size());
 		fflush(stdout);
 	}
-	#endif
 
 	if(*m_mode_send_vertices_sequence_id>(int)m_myReads->size()-1){
 		// flush data
 		flushAll(m_outboxAllocator,m_outbox,rank);
 		if(m_pendingMessages==0){
+			#ifdef ASSERT
+			assert(m_bufferedData.isEmpty());
+			assert(m_bufferedDataForIngoingEdges.isEmpty());
+			assert(m_bufferedDataForOutgoingEdges.isEmpty());
+			assert(m_buffersForOutgoingEdgesToDelete.isEmpty());
+			assert(m_buffersForIngoingEdgesToDelete.isEmpty());
+			#endif
+
 			Message aMessage(NULL,0, MPI_UNSIGNED_LONG_LONG, MASTER_RANK, RAY_MPI_TAG_VERTICES_DISTRIBUTED,rank);
 			m_outbox->push_back(aMessage);
 			*m_mode_send_vertices=false;
@@ -80,109 +99,116 @@ void VerticesExtractor::process(int*m_mode_send_vertices_sequence_id,
 		assert(m_wordSize<=32);
 		#endif
 
-		m_hasPreviousVertex=false;
 
-		for(int p=0;p<lll;p++){
-			memcpy(memory,readSequence+p,m_wordSize);
-			memory[m_wordSize]='\0';
-			if(isValidDNA(memory)){
-				uint64_t a=wordId(memory);
-				int rankToFlush=0;
+		int p=(m_mode_send_vertices_sequence_id_position);
+		memcpy(memory,readSequence+p,m_wordSize);
+		memory[m_wordSize]='\0';
+		if(isValidDNA(memory)){
+			uint64_t a=wordId(memory);
+			int rankToFlush=0;
 
-				rankToFlush=vertexRank(a,size);
-				m_bufferedData.addAt(rankToFlush,a);
+			rankToFlush=vertexRank(a,size);
+			m_bufferedData.addAt(rankToFlush,a);
 
-				if(m_bufferedData.flush(rankToFlush,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,false)){
-					m_pendingMessages++;
-				}
-
-				if(m_hasPreviousVertex){
-					// outgoing edge
-					int outgoingRank=vertexRank(m_previousVertex,size);
-					m_bufferedDataForOutgoingEdges.addAt(outgoingRank,m_previousVertex);
-					m_bufferedDataForOutgoingEdges.addAt(outgoingRank,a);
-
-					if(m_bufferedDataForOutgoingEdges.needsFlushing(outgoingRank,2)){
-						if(m_bufferedData.flush(outgoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
-							m_pendingMessages++;
-						}
-					}
-
-					if(m_bufferedDataForOutgoingEdges.flush(outgoingRank,2,RAY_MPI_TAG_OUT_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
-						m_pendingMessages++;
-					}
-
-					// ingoing edge
-					int ingoingRank=vertexRank(a,size);
-					m_bufferedDataForIngoingEdges.addAt(ingoingRank,m_previousVertex);
-					m_bufferedDataForIngoingEdges.addAt(ingoingRank,a);
-
-					if(m_bufferedDataForIngoingEdges.needsFlushing(ingoingRank,2)){
-						if(m_bufferedData.flush(ingoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
-							m_pendingMessages++;
-						}
-					}
-
-					if(m_bufferedDataForIngoingEdges.flush(ingoingRank,2,RAY_MPI_TAG_IN_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
-						m_pendingMessages++;
-					}
-				}
-
-				// reverse complement
-				uint64_t b=complementVertex(a,m_wordSize,m_colorSpaceMode);
-
-				rankToFlush=vertexRank(b,size);
-				m_bufferedData.addAt(rankToFlush,b);
-
-				if(m_bufferedData.flush(rankToFlush,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,false)){
-					m_pendingMessages++;
-				}
-
-				if(m_hasPreviousVertex){
-					// outgoing edge
-					int outgoingRank=vertexRank(b,size);
-					m_bufferedDataForOutgoingEdges.addAt(outgoingRank,b);
-					m_bufferedDataForOutgoingEdges.addAt(outgoingRank,m_previousVertexRC);
-
-					if(m_bufferedDataForOutgoingEdges.needsFlushing(outgoingRank,2)){
-						if(m_bufferedData.flush(outgoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
-							m_pendingMessages++;
-						}
-					}
-
-					if(m_bufferedDataForOutgoingEdges.flush(outgoingRank,2,RAY_MPI_TAG_OUT_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
-						m_pendingMessages++;
-					}
-
-					// ingoing edge
-					int ingoingRank=vertexRank(m_previousVertexRC,size);
-					m_bufferedDataForIngoingEdges.addAt(ingoingRank,b);
-					m_bufferedDataForIngoingEdges.addAt(ingoingRank,m_previousVertexRC);
-
-					if(m_bufferedDataForIngoingEdges.needsFlushing(ingoingRank,2)){
-						if(m_bufferedData.flush(ingoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
-							m_pendingMessages++;
-						}
-					}
-
-					if(m_bufferedDataForIngoingEdges.flush(ingoingRank,2,RAY_MPI_TAG_IN_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
-						m_pendingMessages++;
-					}
-				}
-
-				// there is a previous vertex.
-				m_hasPreviousVertex=true;
-				m_previousVertex=a;
-				m_previousVertexRC=b;
-			}else{
-				m_hasPreviousVertex=false;
+			if(m_bufferedData.flush(rankToFlush,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,false)){
+				m_pendingMessages++;
 			}
+
+			if(m_hasPreviousVertex){
+				// outgoing edge
+				int outgoingRank=vertexRank(m_previousVertex,size);
+				m_bufferedDataForOutgoingEdges.addAt(outgoingRank,m_previousVertex);
+				m_bufferedDataForOutgoingEdges.addAt(outgoingRank,a);
+
+				if(m_bufferedDataForOutgoingEdges.needsFlushing(outgoingRank,2)){
+					if(m_bufferedData.flush(outgoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
+						m_pendingMessages++;
+					}
+				}
+
+				if(m_bufferedDataForOutgoingEdges.flush(outgoingRank,2,RAY_MPI_TAG_OUT_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
+					m_pendingMessages++;
+				}
+
+				// ingoing edge
+				int ingoingRank=vertexRank(a,size);
+				m_bufferedDataForIngoingEdges.addAt(ingoingRank,m_previousVertex);
+				m_bufferedDataForIngoingEdges.addAt(ingoingRank,a);
+
+				if(m_bufferedDataForIngoingEdges.needsFlushing(ingoingRank,2)){
+					if(m_bufferedData.flush(ingoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
+						m_pendingMessages++;
+					}
+				}
+
+				if(m_bufferedDataForIngoingEdges.flush(ingoingRank,2,RAY_MPI_TAG_IN_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
+					m_pendingMessages++;
+				}
+			}
+
+			// reverse complement
+			uint64_t b=complementVertex(a,m_wordSize,m_colorSpaceMode);
+
+			rankToFlush=vertexRank(b,size);
+			m_bufferedData.addAt(rankToFlush,b);
+
+			if(m_bufferedData.flush(rankToFlush,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,false)){
+				m_pendingMessages++;
+			}
+
+			if(m_hasPreviousVertex){
+				// outgoing edge
+				int outgoingRank=vertexRank(b,size);
+				m_bufferedDataForOutgoingEdges.addAt(outgoingRank,b);
+				m_bufferedDataForOutgoingEdges.addAt(outgoingRank,m_previousVertexRC);
+
+				if(m_bufferedDataForOutgoingEdges.needsFlushing(outgoingRank,2)){
+					if(m_bufferedData.flush(outgoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
+						m_pendingMessages++;
+					}
+				}
+
+				if(m_bufferedDataForOutgoingEdges.flush(outgoingRank,2,RAY_MPI_TAG_OUT_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
+					m_pendingMessages++;
+				}
+
+				// ingoing edge
+				int ingoingRank=vertexRank(m_previousVertexRC,size);
+				m_bufferedDataForIngoingEdges.addAt(ingoingRank,b);
+				m_bufferedDataForIngoingEdges.addAt(ingoingRank,m_previousVertexRC);
+
+				if(m_bufferedDataForIngoingEdges.needsFlushing(ingoingRank,2)){
+					if(m_bufferedData.flush(ingoingRank,1,RAY_MPI_TAG_VERTICES_DATA,m_outboxAllocator,m_outbox,rank,true)){
+						m_pendingMessages++;
+					}
+				}
+
+				if(m_bufferedDataForIngoingEdges.flush(ingoingRank,2,RAY_MPI_TAG_IN_EDGES_DATA,m_outboxAllocator,m_outbox,rank,false)){
+					m_pendingMessages++;
+				}
+			}
+
+			// there is a previous vertex.
+			m_hasPreviousVertex=true;
+			m_previousVertex=a;
+			m_previousVertexRC=b;
+		}else{
+			m_hasPreviousVertex=false;
 		}
-		(*m_mode_send_vertices_sequence_id)++;
+
+		(m_mode_send_vertices_sequence_id_position++);
+		if((m_mode_send_vertices_sequence_id_position)==lll){
+			m_hasPreviousVertex=false;
+			(*m_mode_send_vertices_sequence_id)++;
+			(m_mode_send_vertices_sequence_id_position)=0;
+		}
 	}
 }
 
 void VerticesExtractor::constructor(int size){
+	m_outbox=NULL;
+	m_mustFlushBuffers=false;
+	m_mode_send_vertices_sequence_id_position=0;
 	m_hasPreviousVertex=false;
 	m_bufferedData.constructor(size,MAXIMUM_MESSAGE_SIZE_IN_BYTES);
 	m_bufferedDataForOutgoingEdges.constructor(size,MAXIMUM_MESSAGE_SIZE_IN_BYTES);
@@ -207,18 +233,15 @@ void VerticesExtractor::enableReducer(){
 	m_thresholdForReduction=m_reductionPeriod;
 }
 
-void VerticesExtractor::setReadiness(RingAllocator*outboxAllocator,StaticVector*outbox,int rank){
+void VerticesExtractor::setReadiness(){
 	#ifdef ASSERT
 	assert(m_pendingMessages>0);
 	#endif
 	m_pendingMessages--;
 
-	// flush other buffers
-	if(m_pendingMessages==0&&mustTriggerReduction()){
-		flushAll(outboxAllocator,outbox,rank);
+	if(m_mustFlushBuffers){
+		flushBuffers(m_rank,m_outbox,m_outboxAllocator);
 	}
-
-	checkPendingMessagesForReduction(outbox,rank);
 }
 
 bool VerticesExtractor::mustRunReducer(){
@@ -258,6 +281,12 @@ void VerticesExtractor::resetRanksDoneForReduction(){
 }
 
 void VerticesExtractor::updateThreshold(MyForest*a){
+	#ifdef ASSERT
+	assert(m_buffersForIngoingEdgesToDelete.isEmpty());
+	assert(m_buffersForOutgoingEdgesToDelete.isEmpty());
+	assert(m_pendingMessages==0);
+	#endif
+
 	m_thresholdForReduction=a->size()+m_reductionPeriod;
 }
 
@@ -300,6 +329,7 @@ void VerticesExtractor::assertBuffersAreEmpty(){
 	assert(m_bufferedData.isEmpty());
 	assert(m_bufferedDataForOutgoingEdges.isEmpty());
 	assert(m_bufferedDataForIngoingEdges.isEmpty());
+	assert(m_mode_send_vertices_sequence_id_position==0);
 }
 
 bool VerticesExtractor::mustTriggerReduction(){
@@ -309,14 +339,21 @@ bool VerticesExtractor::mustTriggerReduction(){
 void VerticesExtractor::scheduleReduction(StaticVector*outbox,int rank){
 	m_mustTriggerReduction=true;
 
+	flushAll(m_outboxAllocator,m_outbox,rank);
 	checkPendingMessagesForReduction(outbox,rank);
 }
 
 void VerticesExtractor::checkPendingMessagesForReduction(StaticVector*outbox,int rank){
-	if(m_pendingMessages==0 && mustTriggerReduction()){
+	if(m_pendingMessages==0 && mustTriggerReduction()
+	&&(m_mode_send_vertices_sequence_id_position)==0){// trigger at the beginning of a read, not in the middle.
 		Message aMessage(NULL,0,MPI_UNSIGNED_LONG_LONG,MASTER_RANK,RAY_MPI_TAG_ASK_BEGIN_REDUCTION_REPLY,rank);
+		//cout<<"Source="<<rank<<" Destination="<<MASTER_RANK<<" RAY_MPI_TAG_ASK_BEGIN_REDUCTION_REPLY (meanwhile, freezing)"<<endl;
 		outbox->push_back(aMessage);
 		m_mustTriggerReduction=false;
+		(*m_mode)=RAY_SLAVE_MODE_DO_NOTHING;
+		#ifdef ASSERT
+		assertBuffersAreEmpty();
+		#endif
 	}
 }
 
@@ -339,6 +376,9 @@ bool VerticesExtractor::deleteVertices(vector<uint64_t>*verticesToRemove,MyFores
 	if(!m_deletionsInitiated){
 		m_deletionsInitiated=true;
 		m_deletionIterator=0;
+		#ifdef ASSERT
+		assert(m_bufferedData.isEmpty());
+		#endif
 	}else if(m_deletionIterator<(uint64_t)verticesToRemove->size()){
 		uint64_t vertex=verticesToRemove->at(m_deletionIterator);
 		m_deletionIterator++;
@@ -357,9 +397,15 @@ bool VerticesExtractor::deleteVertices(vector<uint64_t>*verticesToRemove,MyFores
 			m_pendingMessages++;
 		}
 	}else{
-		m_pendingMessages+=m_bufferedData.flushAll(RAY_MPI_TAG_DELETE_VERTEX,m_outboxAllocator,m_outbox,rank);
+		if(!m_bufferedData.isEmpty()){
+			m_pendingMessages+=m_bufferedData.flushAll(RAY_MPI_TAG_DELETE_VERTEX,m_outboxAllocator,m_outbox,rank);
+		}
 
 		if(m_pendingMessages==0){
+			#ifdef ASSERT
+			assert(m_pendingMessages==0);
+			assert(m_bufferedData.isEmpty());
+			#endif
 			return true;
 		}
 	}
@@ -371,8 +417,34 @@ void VerticesExtractor::prepareDeletions(){
 }
 
 void VerticesExtractor::flushBuffers(int rank,StaticVector*m_outbox,RingAllocator*m_outboxAllocator){
-	m_pendingMessages+=m_buffersForIngoingEdgesToDelete.flushAll(RAY_MPI_TAG_DELETE_INGOING_EDGE,m_outboxAllocator,m_outbox,rank);
-	m_pendingMessages+=m_buffersForOutgoingEdgesToDelete.flushAll(RAY_MPI_TAG_DELETE_OUTGOING_EDGE,m_outboxAllocator,m_outbox,rank);
+	m_mustFlushBuffers=true;
+	if(!m_buffersForIngoingEdgesToDelete.isEmpty()){// not empty
+		m_pendingMessages+=m_buffersForIngoingEdgesToDelete.flushAll(RAY_MPI_TAG_DELETE_INGOING_EDGE,m_outboxAllocator,m_outbox,rank);
+		#ifdef ASSERT
+		assert(m_pendingMessages>0); // flushed something
+		#endif
+		return;
+	}
+
+	if(!m_buffersForOutgoingEdgesToDelete.isEmpty()){// not empty
+		m_pendingMessages+=m_buffersForOutgoingEdgesToDelete.flushAll(RAY_MPI_TAG_DELETE_OUTGOING_EDGE,m_outboxAllocator,m_outbox,rank);
+		#ifdef ASSERT
+		assert(m_pendingMessages>0);// flushed something
+		#endif
+		return;
+	}
+
+	if(m_pendingMessages==0){
+		#ifdef ASSERT
+		assert(m_pendingMessages==0);
+		assert(m_buffersForOutgoingEdgesToDelete.isEmpty());
+		assert(m_buffersForIngoingEdgesToDelete.isEmpty());
+		#endif
+
+		Message aMessage(NULL,0,MPI_UNSIGNED_LONG_LONG,MASTER_RANK,RAY_MPI_TAG_UPDATE_THRESHOLD_REPLY,rank);
+		m_outbox->push_back(aMessage);
+		m_mustFlushBuffers=false;
+	}
 }
 
 void VerticesExtractor::incrementPendingMessages(){
